Add a workaround for long pipe path names (#28841)
There's a problem on MacOS where the pipe name may be too long to
use as a Unix domain socket. Unfortunately, there's no way in the
version of CoreFX that we reference to override the path name
used for the named pipe. If the path is too long then the compiler
server spins infinitely trying to connect to a path.

Since we can't use our own path, the best we can do right now is
abort compiler server usage if the path would be too long. This
commit adds a check on Unix machines running on CoreCLR to
detect when the constructed path will be too long, and bail out
if this is the case.