Don't create new Solution snapshots if nothing changed
This just replicates the pattern used in a few other .With* methods
in this type. A better fix would be to update ForkProject to do this
automatically, but there's some complexity there as some calls require
the fork to happen even if nothing did change. This is just a quick fix
as the deeper fix is going to require some time and care.