Relax VB type restrictions on null coalescing If
VB's behavior for If(Nothing, t1) diverges slightly from C#, as a
fallout of how VB handles If(Nothing, 1). C# doesn't let the code
compile, but if the unconstrained type parameter is a value type it
would take the first argument always. VB allows If(Nothing, 1), and it
will pick the second argument to the If. So, the unconstrained T
behavior matches this.