Add analyzer options threading (#27397)
This change introduces a new AnalyzerConfigOptionsProvider on AnalyzerOptions
so that analyzer can get access to the non-diagnostic information
provided in analyzer config files. The AnalyzerConfigOptionsProvider provides
an AnalyzerConfigOptions type containing the non-diagnostic severity properties 
of analyzer config files for a given SyntaxTree. Diagnostic severity continues to
be provided on the syntax tree directly.