Fix temp creation with `in` parameters (#29591)
There's a variety of cases in the compiler where we need to store an
expression into a temporary variable, e.g. to sequence side effects
in the right order. This code was not adjusted to deal with in
parameters, so it exhibited various pathologies.

The first case is the simplest -- sometimes when using the "explicit in"
modifier on an argument, the temp creation routine would simply crash.

The other cases are more subtle and mostly have to do with whether or
not the temp is created as a ref local or regular local. This is
specific to the "implicit in", where the "in" keyword is not mentioned
on the argument. In this case, the semantics for the call should be that
the argument is passed by-ref without storing to a copy unless
absolutely necessary. Absolutely necessary in this case is defined as 1)
the argument is an rvalue or 2) the argument requires a conversion to
the parameter type.

A lot of this knowledge was already present in CodeGen, but not available
in lowering, so I've pulled out the HasHome helper to Binding, as I felt that
was the best place for the common code.

Fixes #29371