Add support for displaying reference kind to Find All References window.
Implementation of the custom column support has been kept extensible, so that each reference item can have customized key-values info, where key corresponds to the name of the custom dynamic column and values are set of one or more values the table entry. Columns become dynamically visible if at least one row in the table has a value for it.

Fixes #24877