Fix SDK path issues
The command line tests are largely failing on CoreClr due to the
inability to implicitly find the SDK directory. On desktop it's possible
to use the directory containing the object assembly. On CoreClr there is
no real equivalent. Fixed this by just writing out mscorlib to a
directory on disk and making that directory explicitly the SDK
directory.