Add fallback to MSBuildProjectLoader when project reference is not backed by metadata reference
In MSBuildWorkspace, there's a fair amount of logic to resolve the metadata references that are passed to the compiler
to project references that are defined on the project. This change adds a fallback in the case that a project reference
can't be matched to one of the metadata references. This is primarily a situation where something is wrong with the project
in it's current configuration. For these cases, MSBuildWorkspace will still attempt to add a project reference but also report
a workspace diagnostic. In addtition, this change removes and reports diagnostics for any unresolved metadata references
passed to the compiler after the reference resolution step occurs. In addition, this change exposed a small bug when
ReferenceOutputAssembly is specified for a project reference, which was caught by an existing test and fixed.