Remove event add and raise code completion options
Remove explicit interface implementation code completion suggestions that are of
methodkind EventAdd and EventRemove. Remove the unit test for checking if
property getters and setters are being offered as a code completion. Instead I
modified the ExplicitInterfaceMemberCompletionContainsOnlyValidValues.
It is redundant to have a method that checks only getters and setters are
missing.