Null check baseDirectory before adding to keyFileSearchPaths (#30383)
* Add test for null base directory in CSharpCommandLineParser.Parse

* Only add baseDirectory and outputDirectory to keyFileSearchPaths if present

* Add VisualBasicCommandLineParser test for Nothing baseDirectory

* Null-check baseDirectory and outputDirectory before adding to keyFileSearchPaths

* Add whitespace baseDir tests

* string.IsNullOrWhiteSpace -> string.IsNullOrEmpty

* Add Debug.Assert to ensure baseDirectory is absolute

CSharpCommandLineArguments.Parse makes the assumption that the
baseDirectory is absolute. This assert makes the assumption clear.

* Remove whitespace baseDirectory test

Because base directories are assumed to be absolute, there's not a
reasonable way to test a whitespace-only base directory. We're keeping
the string.IsNullOrEmpty in case relative base directories are accepted
down the line but for now will be unable to have test coverage for the
whitespace-only case.

* Debug.Assert for absolute baseDirectory in VB command line parser

* Only run Win32ResQuotes on WindowsDesktop

* Fix usage of Parse with empty string baseDirectory