Implement new interfaces for adding files
Implement a new interface to allow csproj/msvbprj to add files.
The existing `ICSharpProjecSite.OnSourceFileAdded` and
`IVbCompilerProject.AddFile` had to go back to the hierarchy to check if
the file had any `<Link>` metadata on it, effectively requiring that
these methods only be called on the main thread. However, this metadata
is readily available to csproj/msvbprj; the new
`IProjectSiteEx.AddFilEx` method allow it to be passed in. This
should make it possible to call this method on a background thread.

Also, the language service supports batching up project changes before
pushing them to the workspace, but does not currently expose a way for
csproj/msvbprj to tell it when a batch of changes is starting or ending.
The `StartBatch` and `EndBatch` methods on `IProjectSiteEx` fill this
need.