Use IVsAsyncFileChangeEx throughout Roslyn
This change is a fairly mechanical change to move over to
IVsAsyncFileChangeEx, without making larger changes to Roslyn. Notably,
Roslyn has always called the file watching service with one request at
time, to avoid starving off the thread pool. Since the entrypoints are
now async, we could start doing more in parallel to to allow the
underlying service to implement things more efficiently.

Also, the implementation for FileChangeTracker is decidedly not
great and still uses synchronous entrypoints in a few places. Since
that part of the code is actively going away once we move the remaining
uses too FileChangeWatcher, we'll just do a cheap approach now.