Fix bug in 'yield return await'
The await expression spiller did not handle the case of a `yield return
await` statement, which it should simply handle by lowering the await
into a spill sequence, then removing the await expression from the
`yield return` statement. This is a simple fix that mostly delegates to
the existing infrastructure.

Fixes #30566