Gracefully handle project systems not giving us a file path
We previously didn't throw if a project system refused to give us a file
path for the project file itself. In my refactoring, this accidentally
switched over to a Marshal.ThrowExceptionForHR, which is now causing us
to crash on at least Visual Studio Tools for Office projects. Put this
back for now.

Ultimately, after this "fix" we're back to the behavior described in
dotnet/rosyn#30429: we won't crash, but we won't have a project
file for the project.