Delete VisualStudioMetadataReference.cs and references to it
This code is dead. It was also terribly confusing; everybody always
expected that VisualStudioMetadataReference inherited from
PortableExecutableReference, but it didn't. The nested .Snapshot type
did.

VisualStudioMetadataReference (the type) still exists as a parent
class of VisualStudioMetadataReference, even though there's nothing
actually in the type now. We can fix that up in a later commit
since that'll be fairly messy.