Update metadata references in response to file changes
This has the same hole that we have with source files right now: we
no longer have the logic which will catch modifications between the
moment we start listening and when we actually get the watcher in place.
Further work will close this hole taking advantage of some perf benefits
from the newer file watching APIs.