Fix up the implementation of IDocumentTrackingService
The previous implementation assumed that it'd get a call from the
old DocumentProvider to know when new files were opened so it could
add it to the list of frames. When I deleted the DocumentProvider, that
hookup wasn't working anymore, and so it would never be tracking
anything at all. Oops.

Rather than couple some types together, the implementation is now
more standalone. I've elected, rather than also listening to RDT events
to simply start tracking a frame the first time it's activated.
The only way I can think a document could be made visible without being
activated that we would care about is on first time load when we
re-open existing documents. If you had a bunch of documents open, and
they were in different frames/windows, we won't consider the non-active
documents to be visible until you first focus on them. I'm not sure this
is really much of a loss; we've often seen that many customers have
files reopened on them that they didn't care about, so this might
actually improve solution open performance since we won't be processing
files they may not care about.

This also brings the service up to a slightly more modern/correct
implementation. This is now a proper MEF part that doesn't assume it's
being created on the UI thread, and it manages it's own shutdown rather
than assuming the package will shut it down in the package unload.
The previous implementation also was a bit strange in that it was a
workspace service, but still had a singleton underlying the whole thing.
The singleton was broken in that different files might be in different
workspaces, but it would always give DocumentIds out for all of the
workspaces, despite it being documented as "only returns for the given
workspace". There's still a singleton doing all the Visual Studio
eventing, but there's now per-workspace logic to correctly filter it.

Unlike the previous implementation, this doesn't really do any cleanup
anymore other than ensure that the async tasks started are joined to
avoid things running after MEF shutdown. Practically, us unsubcribing
events brought no value to anything other than spending extra time
during shutdown, so we'll just leave that as is. It feels "wrong" but
it's the way many are arguing this should be done these days.

Lastly, this now uses async service provider to fetch services. Calls
for data do not block on the service provider having done that work,
since doing so would violate other parts of Roslyn assuming that we
wouldn't block on the UI. I think it's fine, since this service is
very much a tool for optimization only. It still uses JTF internally
on calls that we know are coming in from VS (namely, the .Dispose()) so
that's fine.