Make default namespace available for VB in workspace and implement SetProperty for CPS project.
Right now VB doesn't have the concept of "default namespace". But we conjure one in workspace by assigning the value of the project's root namespace to it. So various feature can choose to use it for their own purpose. In the future, we might consider officially exposing "default namespace" for VB project (e.g. through a <defaultnamespace> msbuild property)