Switch calls to GetProjectState if we don't need the snapshot
These methods are called any time we add a new project and we have a
new output path; this was causing us to realize a Project object for
every project in the solution, which was O(n^2) during solution load.
This was adding a few hundred milliseconds of overhead in the immutable
data structure, which in this particular case we don't need.

Fixes dotnet/roslyn#31306.