Replace AnalyzerConfig.GetAnalyzerConfigOptions() with a new API
GetAnalyzerConfigOptions() was doing two separate things:

1. Parse and process the .editorconfig files.
2. Compute how those files apply to source file paths.

This breaks these out into two steps: there's now a new type
AnalyzerConfigSet that can be created from a set of AnalyzerConfigs.
This does the first step, and you can then ask an AnalyzerConfigSet for
what applies to a specific file, which gives you the result of the
second step. This is intended to better fit the IDE model, where we will
do step 1 and then share that to the different places that lazily do
step 2. It also allows for parallelization of step 2 across files.