Update IEditorConfigStorageLocation to take a Dictionary<string, string>
In practice, everything was either assuming the only keys they cared
about were strings, or they were blindly taking the non-strings and
converting them to strings with .ToString().

This is just a temporary hack to do a lot of mechanical changes in a
single commit -- this will go away in a upcoming commit.