Use release bootstrap compiler
The point of using the bootstrap compiler is validating that we can
build the C# code base with the latest C# compiler. Unfortunately the
debug version of the compiler is significantly slower than the Release
version and it slows down our Debug unit test runs accordingly.

This change causes us to prefer a release version of the bootstrap
compiler. The debug bootstrap compiler is still used in the determinism
leg hence we wont't lose coverage here. This should bring our Release
and Debug desktop tests back to the same run time.