Remove assumption that project names are unique
While refactoring the project system shims I assumed what we were
calling the "unique name" was actually unique. It's not, really --
you can have the same name in different solution folders. We only use
this name for hosting IVsContainedLanguage instances, and it turns out
there we were first filtering by IVsHierarchy and only using the name
to distinguish after that. This returns to that pattern.