Sign reference assemblies in legacy
The design of reference assemblies, no matter whether they are generated
using `/ref` or `/refout` is that they are signed if strong signing
would be done for an implementation assembly.

This was implemented correctly for builder based signing but not for COM
based. This PR fixes the signing to be consistent.