Use PipeOptions.CurrentUserOnly
Now that we are targetting netcoreapp2.1 we can use the
`PipeOptions.CurrentUserOnly` option:

- https://github.com/dotnet/corefx/issues/25427

This is implemented by both CoreClr and Mono. It covers all of the
identity checking to make sure the same user is connected to both ends
of a named pipe. This is mostly just an implementation of what we were
already doing via ACLS and private reflection. But now it can simply be
delegated to the runtime entirely