Discover .editorconfig files during evaluation
Currently we discover .editorconfig files using a target and task that run before `CoreCompile`. On the plus side, this is relatively straightforward and yields (almost) exactly the right set of .editorconfig files that could be used by the compiler--e.g., it will find .editorconfig files that are next to .cs/.vb files produced/found by other targets (so long as they run earlier) and will _not_ pass along any .editorconfig files that are excluded due to `root=true`. It can also pass along to the project systems the (almost) exact set of places where .editorconfig files don't exist but would be interesting if they did. This allows the project systems to place file watchers on those locations and rerun the design-time build if they appear.

The big downside is that this doesn't play well with the caches that we use to limit the number of design-time builds. These generally expect "interesting" project changes to alter something in MSBuild's evaluated model of the project--adding a property modifies a project file or one of its imports; adding or removing a code file modifies the project file or causes a glob to evaluate to a different set of items, etc. But as .editorconfig files don't show up in the evaluated model at all our caches don't notice when one is added or removed.

This leads to weird problems. If you add a .editorconfig before opening a solution the proejct system won't see it at all--it will have been passed cached data. And if you add one _after_ opening the solution, the project system (if it has set up file watchers) may see it and request a new design-time build--but that build will just return cached data that doesn't include the file. Similar problems occur with deleting files.

To better support this scenario, [Microsoft/msbuild #4050](https://github.com/Microsoft/msbuild/pull/4050) adds a new intrinsic item function, `GetPathsOfAllFilesAbove`, that maps from a set of items to all the files of a given name that are found next to those items or above them in the directory tree. This commit uses the new functionalty to map from all `<Compile>` items to the applicable .editorconfig files and add them to the evaluated model.

On the plus side this will help us avoid the caching problem and eliminates a set of MSBuild tasks and targets.

On the negative side, we will no longer be able to find .editorconfig files related to `<Compile>` items produced by tasks/targets. This is a minor consideration as we couldn't guarantee our task would "see" these items, anyway, as it would depend on the relative ordering of the targets during a build.

Also, we no longer get the set of "potential" .editorconfig files to pass to the project system. This isn't great, but the project systems can figure out these locations on their own without too much work.