Fix up logic that is determining and setting contexts for files
The basic mental here is to determine a context for a file, we start
with the Running Document Table and grab the IVsHierarchy for that file.
We then chase it through shared projects until we arrive at final
IVsHierarchy. This should be the IVsHierarchy we were given for the
project. In the case of multitargeting projects, there's one last hop
where we fetch a VSHPROPID_ActiveIntellisenseProjectContext that just
returns the project name instead of a IVsHierarchy. If you squint you
could imagine that as if it returned the IVsHierarchy for that project
except we don't have unique ones.

To set the context, we start with the project and walk those
IVsHierarchy pointers backwards. At each step we can do an operation to
figure out which IVsHierarchy needs its pointer updated, and we point it
to build the chain, moving back to the IVsHierarchy that will end up
in the Running Document Table.

This is (still) leaking IVsHierarchies that we subscribe to, and that
will be fixed next.