Correctly free our IVsHierarchy subscriptions when we're done
Because the code wasn't finished, we were leaking IVsHierarchies since
we had no cleanup logic. We hold onto the IVsHierarchies per document
cookie (although we do share the watchers), so if we close a file we
only unhook from the correct IVsHierarchies rather than having to re-
look at everything.