Remove hack which was trying to avoid a deadlock when projects unload
This was setting a magic thread local property which tracked if we
were currently unloading a project, and if so avoided taking a lock
when the context changed. This was somewhat self-inflicted because
we were trying to change project contexts in the Visual Studio shell
during the middle of unload. We're not doing that anymore, and removing
this also allows us to allow project removal off the UI thread.