Fix regression where renaming a CodeElement didn't update node keys
In a larger refactoring I had commented some code out in
NodeKeyValidation.AddProject. This meant that when we did a refactor
rename of a CodeElement, we wouldn't end up updating node keys for all
the existing CodeElements out there. The expectation is CodeElements
are still valid. Uncommenting the code and bringing it back inline with
the current design makes everything work again.

Fixes dotnet/roslyn#31735.