Work around bug in NamedPipeClientStream.ConnectAsync
`NamedPipeClientStream.ConnectAsync` attempts to be asynchronous by
moving the expensive work (actually connecting to the named pipe) off to
a background thread via `Task.Factory.StartNew(Action,
CancellationToken)`. However, this is the wrong overload to call, and in
some situations is guaranteed to _not_ schedule the new `Task` to a
background thread.

This overload will, by default, use whatever `TaskScheduler` is
associated with the "current" `Task`, if any, and only schedule the work
to at threadpool thread if there is no current `Task` or if that `Task`
is already associated with the threadpool `TaskScheduler`. If, on the
other hand, you're on the UI thread of a GUI application (like Visual
Studio) the new `Task` will also end up scheduled to the UI
thread--exactly the outcome we were trying to avoid.

Once this happens you can easily end up in a situation where a
subsequent blocking call (e.g. `Task.Wait(...)`) blocks the UI thread
while effectively waiting for that `Task` to complete... which it can't
do because the UI thread is blocked! Now you've got a single-threaded
deadlock and your UI hangs.

To work around this issue, here we first force ourselves to a
threadpool thread via `Task.Run` and _then_ call `ConnectAsync`. Thus
the `Task` created by `Task.Factory.StartNew` will also end up on a
threadpool thread.